FileAppender: don't force flush data to file system (behavior change!)
FileAppender will no longer call QFile::flush on every message write
due to it dramatically slowing apps when writing a lot of log records
to slow file system. The downside is some log data may be lost if
application crashes before QFile manages to write all of the buffered
data to file system.

It is still possible to revert to an old behavior calling
setFlushOnWrite(true) on FileAppender.